+++ src/windows-emulator/syscalls/section.cpp
@@ -227,7 +227,7 @@ namespace syscalls
 
         const auto reserve_only = section_entry->allocation_attributes == SEC_RESERVE;
         const auto protection = map_nt_to_emulator_protection(section_entry->section_page_protection);
-        const auto address = c.win_emu.memory.allocate_memory(size, protection, reserve_only);
+        const auto address = c.win_emu.memory.allocate_memory(static_cast<size_t>(size), protection, reserve_only);
 
         if (!reserve_only && !file_data.empty())
         {
